allocate compression output buf instead of splitting from the write buf
This lets us use the full write buffer for plaintext data when compression is enabled, and allows for the possibility of dynamically resizing the write buffer in the future, at the cost of requiring more memory for sessions that use compression.

At the time of this writing, the memory required by the compression codec (109k) is significantly more than what is typically reserved for the compression output buffer (2k by default), so this change is relatively negligible.